Ensures ContentType GUIDs
This is an attempt to work towards further backwards-compatibility for the alias/GUID migration.

We override the `ConvertDbToEditor` methods (in both prevalue-editor and property-editor),
and check if the "icContentTypeGuid" value is available, if not, we attempt to get the GUID, based on the alias.

This commit is reliant on InnerContent's PR, https://github.com/umco/umbraco-inner-content/pull/15